% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHQ-package.R
\docType{package}
\name{KHQ-package}
\alias{KHQ}
\alias{KHQ-package}
\title{KHQ: Methods for Calculating King's Health Questionnaire (KHQ) Scores and KHQ Five Dimensions (KHQ5D) Utility Index Scores}
\description{
The KHQ is a disease-specific, self-administered questionnaire 
  designed specific to assess the impact of Urinary incontinence (UI) on QoL. 
  The questionnaire was developed by Kelleher and collaborators in 1997. It is 
  a simple, acceptable and reliable measure to use in the clinical setting and 
  a research tool that is useful in evaluating UI treatment outcomes. This 
  package calculates scores for each dimension of the KHQ; converts KHQ item 
  scores to KHQ5D scores; and also calculates the utility index of the KHQ5D.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/augustobrusaca/KHQ}
  \item Report bugs at \url{https://github.com/augustobrusaca/KHQ/issues}
}

}
\author{
\strong{Maintainer}: Luiz Augusto Brusaca \email{augustobrusaca@gmail.com} (\href{https://orcid.org/0000-0001-8201-6433}{ORCID})

}
\keyword{internal}
