% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHQ5D.R
\name{KHQ5D}
\alias{KHQ5D}
\title{KHQ5D}
\usage{
KHQ5D(
  scores,
  country,
  type,
  author,
  year,
  source,
  save.xlsx = FALSE,
  filename = NULL,
  sheetName = NULL,
  ignore.invalid = FALSE
)
}
\arguments{
\item{scores}{PARAM_DESCRIPTION}

\item{country}{PARAM_DESCRIPTION}

\item{type}{PARAM_DESCRIPTION}

\item{author}{PARAM_DESCRIPTION}

\item{year}{PARAM_DESCRIPTION}

\item{source}{PARAM_DESCRIPTION}

\item{save.xlsx}{PARAM_DESCRIPTION, Default: FALSE}

\item{filename}{PARAM_DESCRIPTION, Default: NULL}

\item{sheetName}{PARAM_DESCRIPTION, Default: NULL}

\item{ignore.invalid}{logical to indicate whether to ignore items data
with invalid, incomplete or missing data, Default: FALSE}
}
\value{
A data frame with utility index scores of the KHQ5D.
}
\description{
Function used to convert the scores of the five dimensions of
the KHQ5D to a single utility index using country specific value set.
}
\details{
Named vector RL, PL, SL, E and S represent Role limitation,
Physical limitation, Social Limitation, Emotions and Sleep, respectfully.
}
\examples{
\dontrun{
KHQ5D(
  scores = KHQ5D_data,
  country = "UK",
  type = "SG",
  author = "Brazier",
  year = 2008,
  source = "KHQ",
  ignore.invalid = TRUE
)

or

#' KHQ5D(
  scores = KHQ5D_data,
  country = "UK",
  type = "SG",
  author = "Brazier",
  year = 2008,
  source = "KHQ",
  save.xlsx = TRUE,
  filename = "Res_KHQ5D_uti_ind.xlsx",
  sheetName = "Utility_Index",
  ignore.invalid = TRUE
)
}

}
\seealso{
\code{\link[KHQ]{KHQConvKHQ5D}}
}
