% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KHQ_data_Kelleher}
\alias{KHQ_data_Kelleher}
\title{KHQ data}
\format{
This data frame has 30 rows and the following 32 columns:
\describe{
\item{1}{General Health Perception: How would you describe your health at the present?}
\item{2}{Incontinence Impact}
\item{3a}{Role Limitations}
\item{3b}{Role Limitations}
\item{4a}{Physical Limitations}
\item{4b}{Physical Limitations}
\item{4c}{Social Limitations}
\item{4d}{Social Limitations}
\item{5a}{Personal Relationship}
\item{5b}{Personal Relationship}
\item{5c}{Personal Relationship}
\item{6a}{Emotions}
\item{6b}{Emotions}
\item{6c}{Emotions}
\item{7a}{Sleep/Energy}
\item{7b}{Sleep/Energy}
\item{8a}{Severity Measures}
\item{8b}{Severity Measures}
\item{8c}{Severity Measures}
\item{8d}{Severity Measures}
\item{8e}{Severity Measures}
\item{9a}{Symptom Severity Scale}
\item{9b}{Symptom Severity Scale}
\item{9c}{Symptom Severity Scale}
\item{9d}{Symptom Severity Scale}
\item{9e}{Symptom Severity Scale}
\item{9f}{Symptom Severity Scale}
\item{9g}{Symptom Severity Scale}
\item{9h}{Symptom Severity Scale}
\item{9i}{Symptom Severity Scale}
\item{9j}{Symptom Severity Scale}
\item{9k}{Symptom Severity Scale}
}
}
\source{
Example data
}
\usage{
data(KHQ_data_Kelleher)
}
\description{
This data set contains an example of 30 responses of the KHQ.
}
\references{
Kelleher CJ, Cardozo LD, Khullar V, Salvatore S. A new questionnaire
to assess the quality of life of urinary incontinent women. Br J Obstet Gynaecol.
1997 Dec;104(12):1374-9. doi: 10.1111/j.1471-0528.1997.tb11006.x.
\href{https://pubmed.ncbi.nlm.nih.gov/9422015/}{PubMed}
}
\keyword{dataset}
