% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHQ5D.R
\name{KHQ5D}
\alias{KHQ5D}
\title{KHQ5D}
\usage{
KHQ5D(
  scores,
  country,
  type,
  author,
  year,
  source,
  save.xlsx = FALSE,
  filename = NULL,
  sheetName = NULL,
  ignore.invalid = FALSE
)
}
\arguments{
\item{scores}{data.frame with colnames RL, PL, SL, E, and S representing
Role limitation, Physical limitation, Social Limitation, Emotions
and Sleep. Alternatively an KHQ5D score can be provided in five digit
format e.g., 11111, c(11111).}

\item{country}{string specifying the country in which the value set scores
were calculated.}

\item{type}{string specifying method type used in deriving value set scores.}

\item{author}{string specifying the author of the article containing the
value set scores.}

\item{year}{string specifying the year the article containing the value set
scores was published.}

\item{source}{string specifying the version of the King's Health Questionnaire used.}

\item{save.xlsx}{logical to indicate whether or not save the results; Default: FALSE.}

\item{filename}{string specifying the file name if save.xlsx = TRUE;
Default: "Res_KHQ5D_uti_ind.xlsx".}

\item{sheetName}{string specifying the sheet name if save.xlsx = TRUE;
Default: "Utility_Index".}

\item{ignore.invalid}{logical to indicate whether to ignore items data
with invalid, incomplete or missing data; Default: FALSE.}
}
\value{
A data frame with utility index scores of the KHQ5D.
}
\description{
Function used to convert the scores of the five dimensions of
the KHQ5D to a single utility index using country specific value set.
}
\details{
Named vector RL, PL, SL, E and S represent Role limitation,
Physical limitation, Social Limitation, Emotions and Sleep, respectfully.
}
\examples{
\dontrun{
KHQ5D(
  scores = KHQ5D_data,
  country = "UK",
  type = "SG",
  author = "Brazier",
  year = 2008,
  source = "KHQ",
  ignore.invalid = TRUE
)

or

#' KHQ5D(
  scores = KHQ5D_data,
  country = "UK",
  type = "SG",
  author = "Brazier",
  year = 2008,
  source = "KHQ",
  save.xlsx = TRUE,
  filename = "Res_KHQ5D_uti_ind.xlsx",
  sheetName = "Utility_Index",
  ignore.invalid = TRUE
)
}

}
\seealso{
\code{\link[KHQ]{KHQConvKHQ5D}}
}
