% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KHQ_Conv_KHQ5D_data}
\alias{KHQ_Conv_KHQ5D_data}
\title{KHQ data to convert to KHQ5D scores}
\format{
This data frame has 30 rows and the following 9 columns:
\describe{
\item{3a}{Role Limitations}
\item{3b}{Role Limitations}
\item{4a}{Physical Limitations}
\item{4b}{Physical Limitations}
\item{4d}{Social Limitations}
\item{5c}{Personal Relationship}
\item{6a}{Emotions}
\item{6b}{Emotions}
\item{7a}{Sleep/Energy}
}
}
\source{
Example data
}
\usage{
KHQ_Conv_KHQ5D_data
}
\description{
This data set contains an example of 30 responses of the KHQ
items used to convert to KHQ5D scores.
}
\examples{
\dontrun{
KHQConvKHQ5D(
  scores = KHQ_Conv_KHQ5D_data,
  ignore.invalid = TRUE
)

or

KHQConvKHQ5D(
  scores = KHQ_Conv_KHQ5D_data,
  save.xlsx = TRUE,
  filename = "KHQ_conv_KHQ5D.xlsx",
  sheetName = "Scores",
  ignore.invalid = TRUE
)
}
}
\references{
Brazier J, Czoski-Murray C, Roberts J, Brown M, Symonds T, Kelleher C.
Estimation of a preference-based index from a condition-specific measure:
the King's Health Questionnaire. Med Decis Making. 2008 Jan-Feb;28(1):113-26.
doi: 10.1177/0272989X07301820.
\href{https://pubmed.ncbi.nlm.nih.gov/17641139/}{PubMed}
}
\keyword{dataset}
