% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KHQConvKHQ5D.R
\name{KHQConvKHQ5D}
\alias{KHQConvKHQ5D}
\title{KHQConvKHQ5D}
\usage{
KHQConvKHQ5D(
  scores,
  save.xlsx = FALSE,
  filename = NULL,
  sheetName = NULL,
  ignore.invalid = FALSE
)
}
\arguments{
\item{scores}{PARAM_DESCRIPTION}

\item{save.xlsx}{PARAM_DESCRIPTION, Default: FALSE}

\item{filename}{PARAM_DESCRIPTION, Default: NULL}

\item{sheetName}{PARAM_DESCRIPTION, Default: NULL}

\item{ignore.invalid}{logical to indicate whether to ignore items data
with invalid, incomplete or missing data, Default: FALSE}
}
\value{
A data frame with scores of the KHQ five dimensions (KHQ5D).
}
\description{
Function used to convert the KHQ item scores to KHQ5D scores.
}
\details{
Nine items of the KHQ are used to calculate the KHQ5D scores. The
items must be named equal the number in the original questionnaire (3a, 3b,
4a, 4b, 4d, 5c, 6a, 6b, 7a).
}
\examples{
\dontrun{
KHQConvKHQ5D(
  scores = KHQ_data,
  ignore.invalid = TRUE
)

or

KHQConvKHQ5D(
  scores = KHQ_data,
  save.xlsx = TRUE,
  filename = "KHQ_conv_KHQ5D.xlsx",
  sheetName = "Scores",
  ignore.invalid = TRUE
)
}
}
\seealso{
\code{\link[KHQ]{KHQ5D}}
}
